Release 1.16.0 (#140)
* Release 1.15.1

* Admin panel - Memory monitoring - Proper page title

* ApphubCi - Fix project structure

* VillanyAtesz - Buttons for changing margin instead of input

* Notebook - Remove deprecated changeSets

* EliteBase - Store messages

* Release 1.16.0

* EliteBase - Process ApproachSettlement messages

* EliteBase - Process Commodity messages

* EliteBase - Process Docking messages

* EliteBase - Further message handlers

* EliteBase - Process Scan and FSD Jump Journal messages

* EliteBase - Process DockedJournalMessage

* Fix int test

* EliteBase - Process all JournalMessages

* EliteBase - Process all messages

* EliteBase - Store StarSystem coordinates in dedicated columns

* EliteBase - Retry processing messages

* EliteBase - Add indexes

* EliteBase - Unit testing ep. I

* EliteBase - Unit testing ep. II

* EliteBase - Unit testing ep. III

* EliteBase - Unit testing ep. IV

* EliteBase - Unit testing ep. V

* EliteBase - Unit testing ep. VI

* Update Heap size for k8s pods

---------

+++ apphub-lib/apphub-lib-common_util/src/main/java/com/github/saphyra/apphub/lib/common_util/AbstractDao.java
@@ -1,6 +1,7 @@
 package com.github.saphyra.apphub.lib.common_util;
 
 import com.github.saphyra.apphub.lib.common_util.converter.Converter;
+import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.repository.CrudRepository;
@@ -9,7 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-@RequiredArgsConstructor
+@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
 @Slf4j
 public abstract class AbstractDao<ENTITY, DOMAIN, ID, REPOSITORY extends CrudRepository<ENTITY, ID>> {
     protected final Converter<ENTITY, DOMAIN> converter;
@@ -24,6 +25,10 @@ public abstract class AbstractDao<ENTITY, DOMAIN, ID, REPOSITORY extends CrudRep
     }
 
     public void deleteAll(List<DOMAIN> domains) {
+        if (domains.isEmpty()) {
+            return;
+        }
+
         repository.deleteAll(converter.convertDomain(domains));
     }
 
@@ -57,7 +62,11 @@ public abstract class AbstractDao<ENTITY, DOMAIN, ID, REPOSITORY extends CrudRep
         repository.save(converter.convertDomain(domain));
     }
 
-    public void saveAll(List<DOMAIN> iterable) {
-        repository.saveAll(converter.convertDomain(iterable));
+    public void saveAll(List<DOMAIN> domains) {
+        if (domains.isEmpty()) {
+            return;
+        }
+
+        repository.saveAll(converter.convertDomain(domains));
     }
 }
